+++ server/api/v1/system/sys_user.go
@@ -320,7 +320,7 @@ func (b *BaseApi) DeleteUser(c *gin.Context) {
 		return
 	}
 	jwtId := utils.GetUserID(c)
-	if jwtId == uint(reqId.ID) {
+	if jwtId == reqId.ID {
 		response.FailWithMessage("删除失败, 自杀失败", c)
 		return
 	}
