+++ server/model/app/community/app_user.go
@@ -22,7 +22,7 @@ type User struct {
 	Phone       string                `json:"phone" form:"phone" gorm:"column:phone;comment:手机;size:45;"`                     //手机
 	Birthday    *time.Time            `json:"birthday" form:"birthday" gorm:"column:birthday;comment:生日;"`                    //生日
 	Sex         *int                  `json:"sex" form:"sex" gorm:"column:sex;comment:性别;size:10;"`                           //性别
-	AuthorityId uint                  `json:"roleId" form:"roleId" gorm:"column:role_id;default:888;comment:用户角色ID;size:20;"` // 用户角色ID
+	AuthorityId uint64                `json:"roleId" form:"roleId" gorm:"column:role_id;default:888;comment:用户角色ID;size:20;"` // 用户角色ID
 	Authority   system.SysAuthority   `json:"authority" gorm:"foreignKey:Id;references:AuthorityId;comment:用户角色"`             //用户角色
 	Authorities []system.SysAuthority `json:"authorities" gorm:"many2many:sys_user_authority;"`
 	UserExtend  UserExtend            `json:"userExtend" gorm:"foreignKey:ID;references:ID;comment:用户扩展"` //用户扩展
