+++ server/model/system/request/jwt.go
@@ -14,8 +14,8 @@ type CustomClaims struct {
 
 type BaseClaims struct {
 	UUID        uuid.UUID
-	ID          uint
+	ID          uint64
 	Username    string
 	NickName    string
-	AuthorityId uint
+	AuthorityId uint64
 }
