+++ server/model/system/request/sys_authority_btn.go
@@ -2,6 +2,6 @@ package request
 
 type SysAuthorityBtnReq struct {
 	MenuID      uint   `json:"menuID"`
-	AuthorityId uint   `json:"authorityId"`
+	AuthorityId uint64 `json:"authorityId"`
 	Selected    []uint `json:"selected"`
 }
