+++ server/model/system/request/sys_casbin.go
@@ -8,7 +8,7 @@ type CasbinInfo struct {
 
 // Casbin structure for input parameters
 type CasbinInReceive struct {
-	AuthorityId uint         `json:"authorityId"` // 权限id
+	AuthorityId uint64       `json:"authorityId"` // 权限id
 	CasbinInfos []CasbinInfo `json:"casbinInfos"`
 }
