+++ server/model/system/response/sys_authority.go
@@ -8,5 +8,5 @@ type SysAuthorityResponse struct {
 
 type SysAuthorityCopyResponse struct {
 	Authority      system.SysAuthority `json:"authority"`
-	OldAuthorityId uint                `json:"oldAuthorityId"` // 旧角色ID
+	OldAuthorityId uint64              `json:"oldAuthorityId"` // 旧角色ID
 }
