+++ server/model/system/sys_authority_menu.go
@@ -3,7 +3,7 @@ package system
 type SysMenu struct {
 	SysBaseMenu
 	MenuId      string                 `json:"menuId" gorm:"comment:菜单ID"`
-	AuthorityId uint                   `json:"-" gorm:"comment:角色ID"`
+	AuthorityId uint64                 `json:"-" gorm:"comment:角色ID"`
 	Children    []SysMenu              `json:"children" gorm:"-"`
 	Parameters  []SysBaseMenuParameter `json:"parameters" gorm:"foreignKey:SysBaseMenuID;references:MenuId"`
 	Btns        map[string]uint        `json:"btns" gorm:"-"`
