+++ server/model/system/sys_menu_btn.go
@@ -6,5 +6,5 @@ type SysBaseMenuBtn struct {
 	global.GVA_MODEL
 	Name          string `json:"name" gorm:"comment:按钮关键key"`
 	Desc          string `json:"desc" gorm:"按钮备注"`
-	SysBaseMenuID uint   `json:"sysBaseMenuID" gorm:"comment:菜单ID"`
+	SysBaseMenuID uint64 `json:"sysBaseMenuID" gorm:"comment:菜单ID"`
 }
