+++ server/model/system/sys_user_authority.go
@@ -2,8 +2,8 @@ package system
 
 // SysUserAuthority 是 sysUser 和 sysAuthority 的连接表
 type SysUserAuthority struct {
-	SysUserId               uint `gorm:"column:sys_user_id"`
-	SysAuthorityAuthorityId uint `gorm:"column:sys_authority_authority_id"`
+	SysUserId               uint64 `gorm:"column:sys_user_id"`
+	SysAuthorityAuthorityId uint64 `gorm:"column:sys_authority_authority_id"`
 }
 
 func (s *SysUserAuthority) TableName() string {
