+++ server/service/system/sys_api.go
@@ -134,7 +134,7 @@ func (apiService *ApiService) GetAllApis() (apis []system.SysApi, err error) {
 //@param: id float64
 //@return: api model.SysApi, err error
 
-func (apiService *ApiService) GetApiById(id int) (api system.SysApi, err error) {
+func (apiService *ApiService) GetApiById(id uint64) (api system.SysApi, err error) {
 	err = global.GVA_DB.Where("id = ?", id).First(&api).Error
 	return
 }
