+++ server/service/system/sys_auto_code.go
@@ -243,7 +243,7 @@ func makeDictTypes(autoCode *system.AutoCodeStruct) {
 //@param: model.AutoCodeStruct
 //@return: err error
 
-func (autoCodeService *AutoCodeService) CreateTemp(autoCode system.AutoCodeStruct, ids ...uint) (err error) {
+func (autoCodeService *AutoCodeService) CreateTemp(autoCode system.AutoCodeStruct, ids ...uint64) (err error) {
 	makeDictTypes(&autoCode)
 	for i := range autoCode.Fields {
 		if autoCode.Fields[i].FieldType == "time.Time" {
@@ -461,7 +461,7 @@ func (autoCodeService *AutoCodeService) addAutoMoveFile(data *tplData) {
 //@param: a *model.AutoCodeStruct
 //@return: err error
 
-func (autoCodeService *AutoCodeService) AutoCreateApi(a *system.AutoCodeStruct) (ids []uint, err error) {
+func (autoCodeService *AutoCodeService) AutoCreateApi(a *system.AutoCodeStruct) (ids []uint64, err error) {
 	var group system.SysApiGroup
 	group.Name = a.Abbreviation
 	group.Description = a.Abbreviation
