+++ server/service/system/sys_base_menu.go
@@ -16,7 +16,7 @@ type BaseMenuService struct{}
 //@param: id float64
 //@return: err error
 
-func (baseMenuService *BaseMenuService) DeleteBaseMenu(id int) (err error) {
+func (baseMenuService *BaseMenuService) DeleteBaseMenu(id uint64) (err error) {
 	err = global.GVA_DB.Preload("MenuBtn").Preload("Parameters").Where("parent_id = ?", id).First(&system.SysBaseMenu{}).Error
 	if err != nil {
 		var menu system.SysBaseMenu
@@ -119,7 +119,7 @@ func (baseMenuService *BaseMenuService) UpdateBaseMenu(menu system.SysBaseMenu) 
 //@param: id float64
 //@return: menu system.SysBaseMenu, err error
 
-func (baseMenuService *BaseMenuService) GetBaseMenuById(id int) (menu system.SysBaseMenu, err error) {
+func (baseMenuService *BaseMenuService) GetBaseMenuById(id uint64) (menu system.SysBaseMenu, err error) {
 	err = global.GVA_DB.Preload("MenuBtn").Preload("Parameters").Where("id = ?", id).First(&menu).Error
 	return
 }
